#2423 moved logic with sgroup to handleHotkeysOverItem
