+++ packages/ketcher-react/src/script/ui/state/hotkeys.ts
@@ -28,7 +28,10 @@ import { load, onAction, removeStructAction } from './shared'
 import actions from '../action'
 import keyNorm from '../data/convert/keynorm'
 import { isIE } from 'react-device-detect'
-import { handleHotkeyOverItem } from './handleHotkeysOverItem'
+import {
+  handleHotkeyOverItem,
+  handleHotkeyOverSGroup
+} from './handleHotkeysOverItem'
 
 export function initKeydownListener(element) {
   return function (dispatch, getState) {
@@ -101,7 +104,10 @@ function keyHandle(dispatch, state, hotKeys, event) {
       // check if atom is currently hovered over
       // in this case we do not want to activate the corresponding tool
       // and just insert the atom directly
-      if (
+      const selected = editor.selection()
+      if (selected) {
+        handleHotkeyOverSGroup({ selected, atomProps: newAction?.opts, editor })
+      } else if (
         hoveredItem &&
         newAction.tool !== 'select' &&
         newAction.dialog !== 'templates'
